refactor(types): adapt review
