+++ src/module.ts
@@ -13,13 +13,12 @@ import {
   addTypeTemplate,
   updateTemplates,
 } from '@nuxt/kit'
-import type { Nuxt } from '@nuxt/schema'
+import type { Nuxt, NuxtApp } from '@nuxt/schema'
 // import webpack from 'webpack' // eslint-disable-line import/no-named-as-default
 
 import { name, version } from '../package.json'
 import { DEFAULTS, type ModuleOptions } from './runtime/types'
 import { replaceCodePlugin } from './replace'
-import { typesGenerator } from './generator/typesGenerator'
 
 const resolver = createResolver(import.meta.url)
 const runtimeDir = resolver.resolve('./runtime')
@@ -72,15 +71,12 @@ export default defineNuxtModule<ModuleOptions>({
 
     addTypeTemplate({
       filename: 'types/cookie-control.d.ts',
-      getContents: typesGenerator,
+      getContents: getTypeTemplate,
       options: moduleOptions,
       write: true,
     })
 
-    nuxt.hook('builder:watch', async (event, path) => {
-      /**
-       * Update the templates when options are changed in a nuxt.config.ts
-       */
+    nuxt.hook('builder:watch', async (_event, path) => {
       if (path.includes('cookieControl')) {
         updateTemplates({
           filter: (t) => t.filename === 'cookie-control-options.ts',
@@ -144,6 +140,22 @@ const blockIframes = (moduleOptions: ModuleOptions) => {
     })
   }
 }
+
+const getTypeTemplate = (data: {
+  nuxt: Nuxt
+  app: NuxtApp
+  options: ModuleOptions
+}) => `// Generated by ${name}
+
+/**
+* Union of the cookie ids specified in the Nuxt configuration.
+*/
+export type CookieID = ${Object.values(data.options.cookies)
+  .map((cookies) => cookies.map((cookie) => `"${cookie.id}"`).join(' | '))
+  .join(' | ')};
+
+export type CookieIDs = Array<CookieID>;
+`
 
 const loadLocales = async (moduleOptions: ModuleOptions) => {
   const locales = moduleOptions.locales
